broadcaster/requirements.txt
nats-py
aiohttp